+++ server/src/main/java/com/piggy/piggyServer/authLogin/LoginRequest.java
@@ -1,6 +1,5 @@
-package com.piggy.piggyServer.auth;
+package com.piggy.piggyServer.authLogin;
 
-import com.piggy.piggyServer.user.RoleUser;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,11 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class RegisterRequest {
-  String name;
-  String lastName;
-  Integer age;
+public class LoginRequest {
   String email;
   String password;
-  RoleUser roleUser;
 }
