+++ server/src/main/java/com/piggy/piggyServer/authLogin/RegisterRequest.java
@@ -1,5 +1,6 @@
-package com.piggy.piggyServer.auth;
+package com.piggy.piggyServer.authLogin;
 
+import com.piggy.piggyServer.user.RoleUser;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -9,7 +10,11 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class LoginRequest {
+public class RegisterRequest {
+  String name;
+  String lastName;
+  Integer age;
   String email;
   String password;
+  RoleUser roleUser;
 }
